move code relying on generated/protos from shared to internal
shared must be pure
